GAZ-120: Refactor and modularize codebase
